Ship radicle services
As Homebrew doesn't allow for multiple services per Formula, we have to
fake it and have a dummy Formula for the ipfs daemon. If we ever want to
see Radicle in homebrew-core as an official formula we need to address
that issue. Also tried to enocde the dependency of the radicle daemon to
the radicle ipfs daemon in the plist, but can't meaningfully test that
until the path issue we currently facing with the packaged rad
executable. What comes out of this are two services:

$ brew services list
Name               Status  User Plist
radicle-alpha      stopped
radicle-alpha-ipfs stopped